Revert "Merge remote-tracking branch 'upstream/p4sync' into p4sync"
This reverts commit 8c71e6fbe6c6acfb9c4fc4c6e55e10b6b8125984, reversing
changes made to 223ea02f14fe572b7482e0701525830ec2958491.

+++ src/common/sdkint.h
@@ -53,7 +53,7 @@ typedef          float     float32;
 typedef         double     float64;
 
 typedef        float32     f32;
-typedef        float64     f64;
+typedef        float32     f64;
 //-----------------------------------------------------------------------------
 // 8-bit <--> 64-bit wide boolean type
 typedef           int8     b8;
